embeddings_api/requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chex==0.1.3
click==8.1.2
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262147379/work
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cycler==0.11.0
dalle-mini @ git+https://github.com/borisdayma/dalle-mini.git@4381fec4dba9e0b55b5f7a68a2c180088d0c5db8
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dm-tree==0.1.7
docker-pycreds==0.4.0
einops==0.4.1
emoji==1.7.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
Flask==2.1.0
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
flatbuffers==2.0
flax==0.4.1
fonttools==4.33.2
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
google==3.0.0
huggingface-hub==0.5.1
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.3
ipykernel==6.13.0
ipython==8.2.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itsdangerous==2.1.2
jax==0.3.7
jaxlib==0.3.7
jedi==0.18.1
Jinja2==3.1.1
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
msgpack==1.0.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.22.3
openai==0.19.0
openpyxl==3.0.10
opt-einsum==3.3.0
optax==0.1.2
packaging==21.3
pandas==1.4.2
pandas-stubs==1.2.0.61
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.8.1
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.20.1
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments==2.12.0
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
pyzmq==22.3.0
regex==2022.3.15
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016711199/work
sacremoses==0.0.49
scikit-learn==1.1.1
scipy==1.8.0
Send2Trash==1.8.0
sentry-sdk==1.5.10
setproctitle==1.2.3
shortuuid==1.0.8
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.2.0
tenacity==8.0.1
terminado==0.13.3
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toolz==0.11.2
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers @ git+https://github.com/huggingface/transformers.git@72728be3dbca26c70dddc8b724eb2c8d901e97dc
typing_extensions==4.2.0
Unidecode==1.3.4
urllib3==1.26.7
vqgan-jax @ git+https://github.com/patil-suraj/vqgan-jax.git@1be20eee476e5d35c30e4ec3ed12222018af8ce4
wandb==0.12.15
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
widgetsnbextension==3.6.0
zipp==3.8.0
